Refactored fs.FS, added fs.PathException
Nearly all methods in fs.FS have to check whether a given path is valid and
whether the file which the path points to exists. Before this commit a lot
of identical code handling these cases was spread around the class. Now
it's driven using exceptions and it's more DRY. Aspect oriented programming
could be even a better solution to this problem, perhaps.